+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Bug fixes 🐛
 - Fix issue where keystrokes and mouse movements could make it past locks while a passphrase attempt was being processed.
+- Multiple unlock timers could be started under certain conditions.
 
 ### Other Changes 🔄
 - Improve performance.